Revamping the data and owner pointers for fields
This is in preparation for allocating large blocks of memory for
multiple passive scalars. In the previous setup, the ownership pointer
was appended to the end of the data block, meaning that the usual linear
array indexing won't work. Now, ownership is tracked via its own
pointer, at the expense of a double dereference to get to the data
block.